Split index.js into separate parts
